requirement.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
torch==2.1.0
tensorflow==2.14.0
pillow==10.1.0
numpy==1.24.3
bcrypt==4.0.1